+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - Fix issue when trying to remove a todo node.
 - Ensure current content is saved before creating a new folder, note or todo.
 - Fix various todo highlighting issues.
+- Fix focus restoration when using a palette.
 
 ## v0.1.3